Adds users express router. #1
